requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
Django==4.2.3
django-braces==1.15.0
django-guardian==2.4.0
django-jalali-date==1.0.2
django-model-utils==4.3.1
django-multiupload==0.6.1
django-star-ratings==0.9.2
fonttools==4.42.1
idna==3.4
jdatetime==4.1.1
kiwisolver==1.4.5
libsass==0.22.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
